Add libssh2-devel package for gpfdist new feature
gpfdist, Cloudberry's parallel file distribution program,
traditionally required data files to be co-located with the gpfdist
process. This limitation made it cumbersome to load data from remote
servers, often requiring an extra data transfer step.
This commit addresses the issue of gpfdist logs continuously growing
and occupying a large amount of disk space in a persistent working
scenario.

libssh2-devel is introduced as a dependency package for the new feature
in the PR apache/cloudberry#1226.